+++ packages/deepspace/src/docs/runtime/auto-client.tsx
@@ -1,4 +1,0 @@
-import { hydrateDocs } from './client'
-
-hydrateDocs()
-